[WIP] Create an API exceptions file - sort out the imports later
